+++ lib/rast/converters/default_converter.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# Converts string into boolean
+class DefaultConverter
+  def convert(object)
+    return nil if object == 'nil'
+
+    object
+  end
+end